requirements.txt
# ============================================================================
# EuroSAT Land-Use Classification — Dependencies
# ============================================================================
# Core ML
torch>=2.0.0
torchvision>=0.15.0
# Data & Scientific
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
Pillow>=10.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
# Configuration
PyYAML>=6.0
# API & Serving
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6
# Logging & Utilities
tqdm>=4.65.0
# Testing
pytest>=7.4.0
# Linting (dev)
flake8>=6.0.0
